Renamed epoch to slot
