+++ packages/core/slot/epoch_test.go
@@ -1,4 +1,4 @@
-package epoch
+package slot
 
 import (
 	"testing"
